+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `preamble_length`, `set_preamble_length` methods
 - `low_data_rate_optimize`, `set_low_data_rate_optimize` methods
 - `frequency_error_indication` `fei` methods, and `FEI` struct
+- `set_over_current_protection`, `set_ocp` methods and `OverCurrentProtection` struct
 
 ## [0.1.0] - 2026-03-07